+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [EZEZ Changelog](https://ezez.dev/guidelines/changelog/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.28.2] - 2025-09-15
+### Change
+- `Input` now forwards ref
+- `Spacer` converted to stitches
+
 ## [0.28.1] - 2025-08-25
 ### Breaking
 - `Choice` component no longer is generic to avoid TS performance issues